Move all beacon_block fns into the same file
