+++ eth2/types/src/beacon_block.rs
@@ -1,11 +1,9 @@
-use super::{BeaconBlockBody, Eth1Data, Hash256};
+use super::{BeaconBlockBody, ChainSpec, Eth1Data, Hash256, ProposalSignedData};
 use crate::test_utils::TestRandom;
 use bls::Signature;
 use rand::RngCore;
 use serde_derive::Serialize;
 use ssz::{hash, Decodable, DecodeError, Encodable, SszStream, TreeHash};
-
-mod signing;
 
 #[derive(Debug, PartialEq, Clone, Serialize)]
 pub struct BeaconBlock {
@@ -21,6 +19,21 @@ pub struct BeaconBlock {
 impl BeaconBlock {
     pub fn canonical_root(&self) -> Hash256 {
         Hash256::from(&self.hash_tree_root()[..])
+    }
+
+    pub fn proposal_root(&self, spec: &ChainSpec) -> Hash256 {
+        let block_without_signature_root = {
+            let mut block_without_signature = self.clone();
+            block_without_signature.signature = spec.empty_signature.clone();
+            block_without_signature.canonical_root()
+        };
+
+        let proposal = ProposalSignedData {
+            slot: self.slot,
+            shard: spec.beacon_chain_shard_number,
+            block_root: block_without_signature_root,
+        };
+        Hash256::from_slice(&proposal.hash_tree_root()[..])
     }
 }
